requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
asgiref==3.3.4
    # via django
autopep8==1.5.7
    # via -r requirements.in
cachetools==4.2.2
    # via google-auth
certifi==2020.12.5
    # via requests
cffi==1.14.5
    # via google-crc32c
chardet==4.0.0
    # via requests
click==7.1.2
    # via pip-tools
crc32c==2.2
    # via -r requirements.in
django-extensions==3.1.3
    # via -r requirements.in
django==3.2.2
    # via
    #   -r requirements.in
    #   django-extensions
    #   djangorestframework
djangorestframework==3.12.4
    # via -r requirements.in
google-api-core==1.26.3
    # via google-cloud-core
google-auth==1.30.0
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-storage
google-cloud-core==1.6.0
    # via google-cloud-storage
google-cloud-storage==1.38.0
    # via -r requirements.in
google-crc32c==1.1.2
    # via google-resumable-media
google-resumable-media==1.2.0
    # via google-cloud-storage
googleapis-common-protos==1.53.0
    # via google-api-core
idna==2.10
    # via requests
packaging==20.9
    # via google-api-core
pep517==0.10.0
    # via pip-tools
pip-tools==6.1.0
    # via -r requirements.in
protobuf==3.16.0
    # via
    #   google-api-core
    #   googleapis-common-protos
psycopg2==2.8.6
    # via -r requirements.in
pyasn1-modules==0.2.8
    # via google-auth
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pycodestyle==2.7.0
    # via autopep8
pycparser==2.20
    # via cffi
pyparsing==2.4.7
    # via packaging
pytz==2021.1
    # via
    #   django
    #   google-api-core
requests==2.25.1
    # via
    #   google-api-core
    #   google-cloud-storage
rsa==4.7.2
    # via google-auth
six==1.16.0
    # via
    #   google-api-core
    #   google-auth
    #   google-cloud-core
    #   google-resumable-media
    #   protobuf
sqlparse==0.4.1
    # via django
toml==0.10.2
    # via
    #   autopep8
    #   pep517
urllib3==1.26.4
    # via requests

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
